tests/wpt/web-platform-tests/fetch/stale-while-revalidate/fetch.tentative.html

<!DOCTYPE html>
<!---
Tentative test against:
https://github.com/whatwg/fetch/pull/853
-->
<meta charset="utf-8">
<title>Tests Stale While Revalidate is not executed for fetch API</title>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="/common/utils.js"></script>
<script>
promise_test(async (test) => {
  var request_token = token();

  const response = await fetch(`stale-script.py?token=` + request_token);
  const response2 = await fetch(`stale-script.py?token=` + request_token);

  assert_not_equals(response.headers.get('Unique-Id'), response2.headers.get('Unique-Id'));
}, 'Second fetch does not return same response');
</script>
